+++ LibCarla/source/carla/sensor/data/ActorDynamicState.h
@@ -13,6 +13,7 @@
 #include "carla/rpc/VehicleFailureState.h"
 #include "carla/rpc/TrafficLightState.h"
 #include "carla/rpc/VehicleControl.h"
+#include "carla/rpc/VehicleLightState.h"
 #include "carla/rpc/WalkerControl.h"
 
 #include <cstdint>
@@ -65,6 +66,10 @@ namespace detail {
     bool has_traffic_light;
     rpc::ActorId traffic_light_id;
     rpc::VehicleFailureState failure_state;
+    float wheel_rotation[4];
+    float wheel_pitch[4];
+    float wheel_height[4];
+    rpc::VehicleLightState::flag_type light_state;
   };
 #pragma pack(pop)
 
@@ -145,8 +150,9 @@ namespace detail {
 
 #pragma pack(pop)
 
+ // Largest element detail::VehicleData
  static_assert(
-    sizeof(ActorDynamicState) == 119u,
+    sizeof(ActorDynamicState) == 147u,
     "Invalid ActorDynamicState size! "
     "If you modified this class please update the size here, else you may "
     "comment this assert, but your platform may have compatibility issues "